1189---Maximum-Number-of-Balloons/solution1189_test.go
package solution1189

import (
	"fmt"
	"testing"
)

var testCases = []struct {
	Input  string
	Output int
}{
	// Mandatory Test Cases
	{
		Input:  "nlaebolko",
		Output: 1,
	},
	{
		Input:  "loonbalxballpoon",
		Output: 2,
	},
	{
		Input:  "leetcode",
		Output: 0,
	},
	{
		Input:  "balon",
		Output: 0,
	},
	{
		Input:  "ballon",
		Output: 0,
	},
	{
		Input:  "krhizmmgmcrecekgyljqkldocicziihtgpqwbticmvuyznragqoyrukzopfmjhjjxemsxmrsxuqmnkrzhgvtgdgtykhcglurvppvcwhrhrjoislonvvglhdciilduvuiebmffaagxerjeewmtcwmhmtwlxtvlbocczlrppmpjbpnifqtlninyzjtmazxdbzwxthpvrfulvrspycqcghuopjirzoeuqhetnbrcdakilzmklxwudxxhwilasbjjhhfgghogqoofsufysmcqeilaivtmfziumjloewbkjvaahsaaggteppqyuoylgpbdwqubaalfwcqrjeycjbbpifjbpigjdnnswocusuprydgrtxuaojeriigwumlovafxnpibjopjfqzrwemoinmptxddgcszmfprdrichjeqcvikynzigleaajcysusqasqadjemgnyvmzmbcfrttrzonwafrnedglhpudovigwvpimttiketopkvqw",
		Output: 10,
	},
	// Additional my custom cases
}

func Test_maxNumberOfBalloons(t *testing.T) {
	var label string
	for _, tc := range testCases {
		label = fmt.Sprintf("Case: Input: %v Output: %v\n", tc.Input, tc.Output)
		t.Run(label, func(t *testing.T) {
			output := maxNumberOfBalloons(tc.Input)
			if output != tc.Output {
				t.Errorf("Expected output to be %v but we got %v", tc.Output, output)
			}
		})
	}
}

func Benchmark_maxNumberOfBalloons(b *testing.B) {
	for i := 0; i < b.N; i++ {
		_ = maxNumberOfBalloons(testCases[0].Input)
	}
}
